refactor: outsourced inputtypes into userInputs.ts
